requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo
pymongo[srv]==4.6.1
scikit-learn

# -e .
